Attempt to simplify S-pairs.
The J-pair definition on GVW only includes the polynomial whose
signature is inherited, and does not subtract the other polynomial
of the pair, as it is done in S-pair definition. This subtraction
step is actually a reduction step, and in GVW, it is left to be done
during reduction.

Using J-pairs instead of S-pairs has the potential of greatly
simplifying the code, and makes it easier to enforce coefficient
consistency, as defined in GVW, that might well be the cause of
issue #1.