Bugfix (sort of): determinism query was encoded wrong, fixing it.
The query was wrongly encoded as:
(out_a[0] != out_b[0]) AND ... AND (out_a[n] != out_b[n]) → non-deterministc

But the correct encoding is with OR: if any output wire can have 2
different values, it is sufficient for it to be deterministic.

Naive OR means multiplying the differences together into a single huge and
impractcal polynomial that doesn't even fit into memory, so this must be
fixed somehow.